+++ lib/src/function.dart
@@ -6,12 +6,23 @@ import 'package:flutter/widgets.dart' show Widget;
 import 'event.dart';
 import 'store.dart';
 
-/// A function that maps the current state of a store to props.
+/// A function that maps store date to props.
 ///
-/// The type parameter `S` is the type of the state of the [Store].
+/// The type parameter `S` is the type of the [StoreData].
 /// The type parameter `P` is the return type of the function.
-typedef StateToPropsMapper<S, P> = P
-    Function(S state, EventProcessor<S> processor, [String? routeName]);
+typedef StoreDataToPropsMapper<S, P> = P Function(StoreData<S> storeData,
+    [String? routeName]);
+
+/// A function that maps double store data to props.
+///
+/// The type parameter `S1` is the type of the 1st [StoreData].
+/// The type parameter `S2` is the type of the 2nd [StoreData].
+/// The type parameter `P` is the return type of the function.
+typedef DoubleStoreDataToPropsMapper<S1, S2, P> = P Function(
+  StoreData<S1> appStoreData,
+  StoreData<S2> pageStoreData, [
+  String? routeName,
+]);
 
 // A function that builds a Widget from a props parameter.
 ///
@@ -25,8 +36,7 @@ typedef WidgetFromPropsBuilder<P> = Widget Function({
 ///
 /// A registered listener is called by the Store at the end of each dispatch method execution.
 typedef EventListener<S> = void Function(
-  S state,
-  EventProcessor<S> processor,
+  StoreData<S> processor,
   Event<S> event,
   UniqueKey key,
 );