+++ CHANGELOG.md
@@ -58,3 +58,4 @@
 ## 0.5.1
 
 * add hashCode and operator== to action classes and event decorators
+* add StoreData and CompositeCallable